select

+++ src/System/DBManager.cpp
@@ -1,4 +1,5 @@
 #include <filesystem>
+#include <unordered_set>
 #include "fort.hpp"
 
 #include "DBManager.h"
@@ -21,7 +22,12 @@ void DBManager::check_db() {
     if (current_dbname.empty()) throw DBException("Please use a database first");
 }
 
+void DBManager::check_db_empty() {
+    if (!current_dbname.empty()) throw DBException("Please input command \"USE MecuryDB\" first");
+}
+
 string DBManager::create_db(string &name) {
+    if (name == MANAGER_NAME) throw DBException("Invalid database name");
     std::error_code code;
     bool suc = fs::create_directories(db_dir / name, code);
     if (suc) return "Created";
@@ -30,6 +36,8 @@ string DBManager::create_db(string &name) {
 }
 
 string DBManager::drop_db(string &name) {
+    check_db_empty();
+    if (name == MANAGER_NAME) throw DBException("Invalid database name");
     std::error_code code;
     auto suc = fs::remove_all(db_dir / name, code);
     if (suc) return "Removed";
@@ -47,10 +55,15 @@ string DBManager::show_dbs() {
 }
 
 string DBManager::use_db(string &name) {
+    if (name == MANAGER_NAME) {
+        this->schemas.clear();
+        this->current_dbname = "";
+        return string("Back to ") + MANAGER_NAME;
+    }
+
     if (!fs::exists(db_dir / name)) return "Database does not exist";
 
     this->schemas.clear();
-
     this->current_dbname = name;
 
     for (auto e : fs::directory_iterator{db_dir / name}) {
@@ -77,3 +90,72 @@ string DBManager::show_indexes() {
     return "Not supported yet";
 }
 
+string DBManager::create_table(Schema &schema) {
+    check_db();
+    // check schema
+    unordered_set<string> column_names;
+    for (auto &column : schema.columns) {
+        if (column_names.find(column.name) == column_names.end())
+            column_names.insert(column.name);
+        else
+            return "Duplicate column name: " + column.name;
+		if(column.has_default){
+			if(column.default_value.type == NULL_TYPE){
+				if(column.not_null) return "Default is NULL but NOT NULL is enabled";
+			}else{
+				if(column.type != column.default_value.type) return "Default value type and field type of column '" + column.name + "' are not identical";
+			}
+		}
+    }
+    for (auto &pk_column : schema.pk.pks) {
+        if (column_names.find(pk_column) == column_names.end()) return "Primary key field not declared: " + pk_column;
+    }
+    for (auto &fk : schema.fks) {
+        for (auto &fk_column : fk.fks) {
+            if (column_names.find(fk_column) == column_names.end()) return "Foreign key field not declared: " + fk_column;
+        }
+        if (schemas.find(fk.ref_table) == schemas.end()) return "Foreign key ref table not found: " + fk.ref_table;
+        Schema &ref_table_schema = schemas[fk.ref_table]; 
+
+        for (auto &fk_ref_col : fk.ref_fks) {
+            if (ref_table_schema.find_column(fk_ref_col) == ref_table_schema.columns.size())
+                return "Foreign key ref field not found: " + fk_ref_col;
+        }
+    }
+    // create directory
+    std::error_code code;
+    bool suc = fs::create_directories(db_dir / current_dbname / schema.table_name, code);
+    if (!suc) {
+        if (code.value() == 0) return "Table already exists";
+        return code.message();
+    }
+    // write
+    suc = schema.write(current_dbname);
+    if (!suc) return "Writing failed";
+
+    this->schemas[schema.table_name] = schema;
+
+    record_handler->createFile((file_name(schema) + ".data").data(), schema.record_type());
+    FileSystem::save();
+
+    return "Created";
+}
+
+string DBManager::drop_table(string name){
+    check_db();
+	// check table
+	auto dir = db_dir / current_dbname / name;
+	std::error_code code;
+	auto suc = fs::remove_all(dir, code);
+	if(suc) return "Removed";
+	if(code.value() == 0) return "Table does not exist";
+
+    schemas.erase(schemas.find(name));
+
+	return code.message();
+}
+
+string DBManager::describe_table(string name){
+    check_db();
+    return get_schema(name).to_str();
+}
